requirements.txt
# Dynamic Cloud Service Composition System - Backend Requirements
flask
flask-cors
firebase-admin
pyrebase4
python-dotenv
numpy
pandas
openpyxl
pytest
google-generativeai
apscheduler
requests
